[SPARC64]: Compat syscall overhaul. 1) Make syscall entry zero-extend all arguments. 2) Sign extend those needed in sys32.S 3) Kill the A() AA() macros, replace with compat_ptr() et al.

+++ arch/sparc64/kernel/signal32.c
@@ -1349,7 +1349,7 @@ out:
 	return ret;
 }
 
-asmlinkage int do_sys32_sigaltstack(u32 ussa, u32 uossa, unsigned long sp)
+asmlinkage long do_sys32_sigaltstack(u32 ussa, u32 uossa, unsigned long sp)
 {
 	stack_t uss, uoss;
 	u32 u_ss_sp = 0;
